Phase 5 완료: MediaPipe Manager v2 기능을 표준 파일로 통합
