+++ systems/mediapipe_manager.py
@@ -1,213 +1,509 @@
+"""
+S-Class DMS v19+ 차세대 MediaPipe Tasks Manager
+최신 MediaPipe Tasks API (0.10.9+) 적용
+"""
+
 import cv2
 import mediapipe as mp
 import time
+import asyncio
 import logging
 import threading
 import queue
-import asyncio
 import numpy as np
-from typing import List
+from typing import List, Dict, Optional, Callable, Any
+from pathlib import Path
+from dataclasses import dataclass
+from enum import Enum
+from collections import deque
+import time
+import cv2
+import mediapipe as mp
+# MediaPipe Tasks API imports
 from mediapipe.tasks import python
-from mediapipe.tasks.python import vision
-from systems.performance import PerformanceOptimizer
+from mediapipe.tasks.python import vision, audio, text
 
 logger = logging.getLogger(__name__)
 
-class EnhancedMediaPipeManager:
-    """향상된 MediaPipe 관리자 (큐 및 스레드 기반 콜백 처리)"""
+class TaskType(Enum):
+    """MediaPipe Tasks 타입"""
+    FACE_LANDMARKER = "face_landmarker"
+    POSE_LANDMARKER = "pose_landmarker" 
+    HAND_LANDMARKER = "hand_landmarker"
+    GESTURE_RECOGNIZER = "gesture_recognizer"
+    OBJECT_DETECTOR = "object_detector"
+    IMAGE_CLASSIFIER = "image_classifier"
+    FACE_DETECTOR = "face_detector"
+    HOLISTIC_LANDMARKER = "holistic_landmarker"
 
-    def __init__(self, analysis_engine):
+@dataclass
+class TaskConfig:
+    """MediaPipe Task 설정"""
+    task_type: TaskType
+    model_path: str
+    max_results: int = 1
+    score_threshold: float = 0.5
+    running_mode: vision.RunningMode = vision.RunningMode.LIVE_STREAM
+    enable_face_blendshapes: bool = False
+    enable_facial_transformation_matrix: bool = False
+    enable_segmentation_masks: bool = False
+    num_poses: int = 1
+    num_hands: int = 2
+    num_faces: int = 1
+    min_detection_confidence: float = 0.5
+    min_tracking_confidence: float = 0.5
+
+class AdvancedMediaPipeManager:
+    """
+    차세대 MediaPipe Tasks Manager - 성능 최적화
+    - 최신 Tasks API 완전 활용
+    - 동적 모델 로딩/언로딩
+    - 성능 최적화 및 메모리 관리
+    - 포괄적 오류 처리
+    """
+    
+    def __init__(self, analysis_engine=None, config_file: Optional[str] = None):
         self.analysis_engine = analysis_engine
-        self.fps_avg_frame_count = 10
-        self.counter, self.start_time, self.current_fps = 0, time.time(), 0.0
-        self.face_landmarker, self.pose_landmarker, self.hand_landmarker, self.object_detector = None, None, None, None
-        self.task_health = {"face": False, "pose": False, "hand": False, "object": False}
-        self.active_tasks = ["face", "pose", "hand", "object"]
-        self.performance_optimizer = PerformanceOptimizer()
-        self.last_timestamp = -1
+        self.active_tasks: Dict[TaskType, Any] = {}
+        self.task_configs: Dict[TaskType, TaskConfig] = {}
+        self.result_callbacks: Dict[TaskType, Callable] = {}
+        self.task_health: Dict[TaskType, bool] = {}
         
-        # 결과 저장을 위한 속성
-        self.last_face_result = None
-        self.last_pose_result = None
-        self.last_hand_result = None
-        self.last_object_result = None
-
-        self.result_queue = queue.Queue()
-        self._shutdown_requested = False  # Initialize shutdown flag
-        self._initialize_tasks()
+        # 성능 모니터링 - 최적화된 deque 사용
+        self.fps_counter = 0
+        self.fps_start_time = time.time()
+        self.current_fps = 0.0
+        self.frame_processing_times = deque(maxlen=100)  # O(1) operations
         
-        self.callback_thread = threading.Thread(target=self._process_callbacks, daemon=True)
-        self.callback_thread.start()
+        # 결과 관리
+        self.latest_results: Dict[TaskType, Any] = {}
+        self.result_queue = queue.Queue(maxsize=100)
+        self.processing_lock = asyncio.Lock()
+        self.last_timestamp = 0
+        
+        # 모델 경로 설정
+        self.model_base_path = Path("models")
+        self.ensure_model_directory()
+        
+        # 비동기 처리 스레드
+        self.callback_thread = None
+        self.running = False
+        
+        # 기본 설정 로드
+        self._load_default_configs()
+        
+        # 외부 설정 파일 로드 (있는 경우)
+        if config_file:
+            self._load_config_file(config_file)
+        
+        logger.info("AdvancedMediaPipeManager v2.0 (최적화) 초기화 완료")
 
-    def _calculate_fps(self):
-        if self.counter > 0 and self.counter % self.fps_avg_frame_count == 0:
-            self.current_fps = self.fps_avg_frame_count / (time.time() - self.start_time)
-            self.start_time = time.time()
-        self.counter += 1
+    def ensure_model_directory(self):
+        """모델 디렉토리 존재 확인"""
+        self.model_base_path.mkdir(exist_ok=True)
+        logger.info(f"모델 디렉토리: {self.model_base_path.absolute()}")
+
+    def _load_default_configs(self):
+        """기본 Task 설정 로드"""
+        
+        # Face Landmarker 설정
+        self.task_configs[TaskType.FACE_LANDMARKER] = TaskConfig(
+            task_type=TaskType.FACE_LANDMARKER,
+            model_path=str(self.model_base_path / "face_landmarker_v2_with_blendshapes.task"),
+            num_faces=1,
+            min_detection_confidence=0.5,
+            min_tracking_confidence=0.5,
+            enable_face_blendshapes=True,
+            enable_facial_transformation_matrix=True
+        )
+        
+        # Pose Landmarker 설정
+        self.task_configs[TaskType.POSE_LANDMARKER] = TaskConfig(
+            task_type=TaskType.POSE_LANDMARKER,
+            model_path=str(self.model_base_path / "pose_landmarker_heavy.task"),
+            num_poses=1,
+            min_detection_confidence=0.5,
+            min_tracking_confidence=0.5,
+            enable_segmentation_masks=True
+        )
+        
+        # Hand Landmarker 설정
+        self.task_configs[TaskType.HAND_LANDMARKER] = TaskConfig(
+            task_type=TaskType.HAND_LANDMARKER,
+            model_path=str(self.model_base_path / "hand_landmarker.task"),
+            num_hands=2,
+            min_detection_confidence=0.7,
+            min_tracking_confidence=0.5
+        )
+        
+        # Gesture Recognizer 설정 (새로운 기능)
+        self.task_configs[TaskType.GESTURE_RECOGNIZER] = TaskConfig(
+            task_type=TaskType.GESTURE_RECOGNIZER,
+            model_path=str(self.model_base_path / "gesture_recognizer.task"),
+            num_hands=2,
+            min_detection_confidence=0.7,
+            min_tracking_confidence=0.5
+        )
+        
+        # Object Detector 설정
+        self.task_configs[TaskType.OBJECT_DETECTOR] = TaskConfig(
+            task_type=TaskType.OBJECT_DETECTOR,
+            model_path=str(self.model_base_path / "efficientdet_lite0.tflite"),
+            max_results=5,
+            score_threshold=0.3
+        )
+        
+        # Holistic Landmarker 설정 (최신 통합 모델)
+        self.task_configs[TaskType.HOLISTIC_LANDMARKER] = TaskConfig(
+            task_type=TaskType.HOLISTIC_LANDMARKER,
+            model_path=str(self.model_base_path / "holistic_landmarker.task"),
+            min_detection_confidence=0.5,
+            min_tracking_confidence=0.5
+        )
+
+    def _load_config_file(self, config_file: str):
+        """외부 설정 파일 로드"""
+        try:
+            # JSON 또는 YAML 설정 파일 파싱 로직
+            logger.info(f"설정 파일 로드: {config_file}")
+            # 구현 필요 시 추가
+        except Exception as e:
+            logger.warning(f"설정 파일 로드 실패: {e}")
+
+    async def initialize_task(self, task_type: TaskType) -> bool:
+        """개별 Task 초기화"""
+        if task_type in self.active_tasks:
+            logger.warning(f"{task_type.value} 이미 초기화됨")
+            return True
+        
+        config = self.task_configs.get(task_type)
+        if not config:
+            logger.error(f"{task_type.value} 설정 없음")
+            return False
+        
+        try:
+            # 모델 파일 존재 확인
+            if not Path(config.model_path).exists():
+                logger.warning(f"모델 파일 없음: {config.model_path}")
+                return False
+            
+            # Base options 설정
+            base_options = python.BaseOptions(
+                model_asset_path=config.model_path
+            )
+            
+            # Task별 초기화
+            if task_type == TaskType.FACE_LANDMARKER:
+                task = await self._initialize_face_landmarker(base_options, config)
+            elif task_type == TaskType.POSE_LANDMARKER:
+                task = await self._initialize_pose_landmarker(base_options, config)
+            elif task_type == TaskType.HAND_LANDMARKER:
+                task = await self._initialize_hand_landmarker(base_options, config)
+            elif task_type == TaskType.GESTURE_RECOGNIZER:
+                task = await self._initialize_gesture_recognizer(base_options, config)
+            elif task_type == TaskType.OBJECT_DETECTOR:
+                task = await self._initialize_object_detector(base_options, config)
+            elif task_type == TaskType.HOLISTIC_LANDMARKER:
+                task = await self._initialize_holistic_landmarker(base_options, config)
+            else:
+                logger.error(f"지원되지 않는 Task 타입: {task_type}")
+                return False
+            
+            if task:
+                self.active_tasks[task_type] = task
+                self.task_health[task_type] = True
+                logger.info(f"✅ {task_type.value} 초기화 완료")
+                return True
+                
+        except Exception as e:
+            logger.error(f"❌ {task_type.value} 초기화 실패: {e}")
+            self.task_health[task_type] = False
+            
+        return False
+
+    async def _initialize_face_landmarker(self, base_options, config):
+        """Face Landmarker 초기화"""
+        options = vision.FaceLandmarkerOptions(
+            base_options=base_options,
+            running_mode=config.running_mode,
+            num_faces=config.num_faces,
+            min_detection_confidence=config.min_detection_confidence,
+            min_tracking_confidence=config.min_tracking_confidence,
+            output_face_blendshapes=config.enable_face_blendshapes,
+            output_facial_transformation_matrixes=config.enable_facial_transformation_matrix,
+            result_callback=self._create_result_callback(TaskType.FACE_LANDMARKER)
+        )
+        return vision.FaceLandmarker.create_from_options(options)
+
+    async def _initialize_pose_landmarker(self, base_options, config):
+        """Pose Landmarker 초기화"""
+        options = vision.PoseLandmarkerOptions(
+            base_options=base_options,
+            running_mode=config.running_mode,
+            num_poses=config.num_poses,
+            min_detection_confidence=config.min_detection_confidence,
+            min_tracking_confidence=config.min_tracking_confidence,
+            output_segmentation_masks=config.enable_segmentation_masks,
+            result_callback=self._create_result_callback(TaskType.POSE_LANDMARKER)
+        )
+        return vision.PoseLandmarker.create_from_options(options)
+
+    async def _initialize_hand_landmarker(self, base_options, config):
+        """Hand Landmarker 초기화"""
+        options = vision.HandLandmarkerOptions(
+            base_options=base_options,
+            running_mode=config.running_mode,
+            num_hands=config.num_hands,
+            min_detection_confidence=config.min_detection_confidence,
+            min_tracking_confidence=config.min_tracking_confidence,
+            result_callback=self._create_result_callback(TaskType.HAND_LANDMARKER)
+        )
+        return vision.HandLandmarker.create_from_options(options)
+
+    async def _initialize_gesture_recognizer(self, base_options, config):
+        """Gesture Recognizer 초기화"""
+        options = vision.GestureRecognizerOptions(
+            base_options=base_options,
+            running_mode=config.running_mode,
+            num_hands=config.num_hands,
+            min_detection_confidence=config.min_detection_confidence,
+            min_tracking_confidence=config.min_tracking_confidence,
+            result_callback=self._create_result_callback(TaskType.GESTURE_RECOGNIZER)
+        )
+        return vision.GestureRecognizer.create_from_options(options)
+
+    async def _initialize_object_detector(self, base_options, config):
+        """Object Detector 초기화"""
+        options = vision.ObjectDetectorOptions(
+            base_options=base_options,
+            running_mode=config.running_mode,
+            max_results=config.max_results,
+            score_threshold=config.score_threshold,
+            result_callback=self._create_result_callback(TaskType.OBJECT_DETECTOR)
+        )
+        return vision.ObjectDetector.create_from_options(options)
+
+    async def _initialize_holistic_landmarker(self, base_options, config):
+        """Holistic Landmarker 초기화 (최신 통합 모델)"""
+        options = vision.HolisticLandmarkerOptions(
+            base_options=base_options,
+            running_mode=config.running_mode,
+            min_detection_confidence=config.min_detection_confidence,
+            min_tracking_confidence=config.min_tracking_confidence,
+            result_callback=self._create_result_callback(TaskType.HOLISTIC_LANDMARKER)
+        )
+        return vision.HolisticLandmarker.create_from_options(options)
+
+    def _create_result_callback(self, task_type: TaskType):
+        """결과 콜백 함수 생성"""
+        def callback(result, output_image, timestamp_ms):
+            try:
+                self.latest_results[task_type] = result
+                self.result_queue.put((task_type, result, timestamp_ms), timeout=0.1)
+            except queue.Full:
+                logger.warning(f"{task_type.value} 결과 큐 오버플로우")
+            except Exception as e:
+                logger.error(f"{task_type.value} 콜백 오류: {e}")
+        return callback
+
+    async def initialize_all_tasks(self) -> Dict[TaskType, bool]:
+        """모든 설정된 Task 초기화"""
+        results = {}
+        
+        # 기본 Tasks 초기화
+        core_tasks = [
+            TaskType.FACE_LANDMARKER,
+            TaskType.POSE_LANDMARKER,
+            TaskType.HAND_LANDMARKER
+        ]
+        
+        # 선택적 Tasks
+        optional_tasks = [
+            TaskType.GESTURE_RECOGNIZER,
+            TaskType.OBJECT_DETECTOR,
+            TaskType.HOLISTIC_LANDMARKER
+        ]
+        
+        # 병렬 초기화
+        for task_type in core_tasks:
+            results[task_type] = await self.initialize_task(task_type)
+        
+        # 선택적 Tasks (실패해도 계속)
+        for task_type in optional_tasks:
+            try:
+                results[task_type] = await self.initialize_task(task_type)
+            except Exception as e:
+                logger.warning(f"선택적 Task {task_type.value} 초기화 실패: {e}")
+                results[task_type] = False
+        
+        # 콜백 처리 스레드 시작
+        if not self.callback_thread or not self.callback_thread.is_alive():
+            self.running = True
+            self.callback_thread = threading.Thread(target=self._process_callbacks, daemon=True)
+            self.callback_thread.start()
+        
+        initialized_count = sum(results.values())
+        total_count = len(results)
+        logger.info(f"Task 초기화 완료: {initialized_count}/{total_count}")
+        
+        return results
 
     def _process_callbacks(self):
+        """비동기 콜백 처리"""
         loop = asyncio.new_event_loop()
         asyncio.set_event_loop(loop)
         
-        # Add shutdown flag for safer exit
-        self._shutdown_requested = False
+        logger.info("콜백 처리 스레드 시작")
         
-        while not self._shutdown_requested:
+        while self.running:
             try:
-                # Add timeout to prevent infinite blocking
-                result_type, result, timestamp = self.result_queue.get(timeout=1.0)
-                if result_type == 'shutdown':
-                    self._shutdown_requested = True
+                task_type, result, timestamp = self.result_queue.get(timeout=1.0)
+                
+                if task_type == "shutdown":
                     break
                 
+                # Analysis engine으로 결과 전달
                 if self.analysis_engine:
-                    if result_type == 'face' and hasattr(self.analysis_engine, 'on_face_result'):
-                        loop.run_until_complete(self.analysis_engine.on_face_result(result, timestamp=timestamp))
-                    elif result_type == 'pose' and hasattr(self.analysis_engine, 'on_pose_result'):
-                        loop.run_until_complete(self.analysis_engine.on_pose_result(result, timestamp=timestamp))
-                    elif result_type == 'hand' and hasattr(self.analysis_engine, 'on_hand_result'):
-                        loop.run_until_complete(self.analysis_engine.on_hand_result(result, timestamp=timestamp))
-                    elif result_type == 'object' and hasattr(self.analysis_engine, 'on_object_result'):
-                        loop.run_until_complete(self.analysis_engine.on_object_result(result, timestamp=timestamp))
+                    loop.run_until_complete(self._forward_result_to_analysis_engine(task_type, result, timestamp))
+                    
             except queue.Empty:
-                # Timeout occurred, check if shutdown was requested
                 continue
             except Exception as e:
-                logger.error(f"Callback 처리 중 오류: {e}")
-                # On critical errors, consider shutting down gracefully
-                if isinstance(e, (KeyboardInterrupt, SystemExit)):
-                    self._shutdown_requested = True
-                    break
+                logger.error(f"콜백 처리 오류: {e}")
         
-        logger.info("Callback processing loop exiting gracefully")
         loop.close()
+        logger.info("콜백 처리 스레드 종료")
 
-    def _on_face_result(self, result, output_image, timestamp_ms):
-        self.last_face_result = result
-        self.result_queue.put(('face', result, timestamp_ms))
-
-    def _on_pose_result(self, result, output_image, timestamp_ms):
-        self.last_pose_result = result
-        self.result_queue.put(('pose', result, timestamp_ms))
-
-    def _on_hand_result(self, result, output_image, timestamp_ms):
-        self.last_hand_result = result
-        self.result_queue.put(('hand', result, timestamp_ms))
-
-    def _on_object_result(self, result, output_image, timestamp_ms):
-        self.last_object_result = result
-        self.result_queue.put(('object', result, timestamp_ms))
-
-    def _initialize_tasks(self):
-        # FaceLandmarker
+    async def _forward_result_to_analysis_engine(self, task_type: TaskType, result, timestamp):
+        """Analysis Engine으로 결과 전달"""
         try:
-            face_options = vision.FaceLandmarkerOptions(
-                base_options=python.BaseOptions(model_asset_path='models/face_landmarker.task'),
-                running_mode=vision.RunningMode.LIVE_STREAM,
-                num_faces=1,
-                output_face_blendshapes=True,
-                output_facial_transformation_matrixes=True,
-                result_callback=self._on_face_result)
-            self.face_landmarker = vision.FaceLandmarker.create_from_options(face_options)
-            self.task_health["face"] = True
-            logger.info("✅ FaceLandmarker 초기화 완료")
+            if hasattr(self.analysis_engine, 'on_face_result') and task_type == TaskType.FACE_LANDMARKER:
+                await self.analysis_engine.on_face_result(result, timestamp=timestamp)
+            elif hasattr(self.analysis_engine, 'on_pose_result') and task_type == TaskType.POSE_LANDMARKER:
+                await self.analysis_engine.on_pose_result(result, timestamp=timestamp)
+            elif hasattr(self.analysis_engine, 'on_hand_result') and task_type == TaskType.HAND_LANDMARKER:
+                await self.analysis_engine.on_hand_result(result, timestamp=timestamp)
+            elif hasattr(self.analysis_engine, 'on_gesture_result') and task_type == TaskType.GESTURE_RECOGNIZER:
+                await self.analysis_engine.on_gesture_result(result, timestamp=timestamp)
+            elif hasattr(self.analysis_engine, 'on_object_result') and task_type == TaskType.OBJECT_DETECTOR:
+                await self.analysis_engine.on_object_result(result, timestamp=timestamp)
+            elif hasattr(self.analysis_engine, 'on_holistic_result') and task_type == TaskType.HOLISTIC_LANDMARKER:
+                await self.analysis_engine.on_holistic_result(result, timestamp=timestamp)
         except Exception as e:
-            logger.error(f"❌ FaceLandmarker 초기화 실패: {e}")
+            logger.error(f"Analysis engine 전달 오류 ({task_type.value}): {e}")
 
-        # PoseLandmarker
-        try:
-            pose_options = vision.PoseLandmarkerOptions(
-                base_options=python.BaseOptions(model_asset_path='models/pose_landmarker_full.task'),
-                running_mode=vision.RunningMode.LIVE_STREAM,
-                num_poses=1,
-                result_callback=self._on_pose_result)
-            self.pose_landmarker = vision.PoseLandmarker.create_from_options(pose_options)
-            self.task_health["pose"] = True
-            logger.info("✅ PoseLandmarker 초기화 완료")
-        except Exception as e:
-            logger.error(f"❌ PoseLandmarker 초기화 실패: {e}")
-
-        # HandLandmarker
-        try:
-            hand_options = vision.HandLandmarkerOptions(
-                base_options=python.BaseOptions(model_asset_path='models/hand_landmarker.task'),
-                running_mode=vision.RunningMode.LIVE_STREAM,
-                num_hands=2,
-                result_callback=self._on_hand_result)
-            self.hand_landmarker = vision.HandLandmarker.create_from_options(hand_options)
-            self.task_health["hand"] = True
-            logger.info("✅ HandLandmarker 초기화 완료")
-        except Exception as e:
-            logger.error(f"❌ HandLandmarker 초기화 실패: {e}")
-
-        # ObjectDetector
-        try:
-            object_options = vision.ObjectDetectorOptions(
-                base_options=python.BaseOptions(model_asset_path='models/efficientdet_lite0.tflite'),
-                running_mode=vision.RunningMode.LIVE_STREAM,
-                max_results=5,
-                score_threshold=0.3,
-                result_callback=self._on_object_result)
-            self.object_detector = vision.ObjectDetector.create_from_options(object_options)
-            self.task_health["object"] = True
-            logger.info("✅ ObjectDetector 초기화 완료")
-        except Exception as e:
-            logger.error(f"❌ ObjectDetector 초기화 실패: {e}")
-
-    def run_tasks(self, frame):
+    async def process_frame(self, frame: np.ndarray) -> Dict[TaskType, Any]:
+        """
+        프레임 처리 - 모든 활성 Task에서 동시 처리
+        """
+        start_time = time.time()
+        
+        # FPS 계산
         self._calculate_fps()
+        
+        # 타임스탬프 생성
         timestamp_ms = int(time.time() * 1000)
         if timestamp_ms <= self.last_timestamp:
             timestamp_ms = self.last_timestamp + 1
         self.last_timestamp = timestamp_ms
-
+        
         try:
+            # BGR to RGB 변환
             rgb_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
             mp_image = mp.Image(image_format=mp.ImageFormat.SRGB, data=rgb_frame)
             
-            if "face" in self.active_tasks and self.face_landmarker:
-                self.face_landmarker.detect_async(mp_image, timestamp_ms)
-            if "pose" in self.active_tasks and self.pose_landmarker:
-                self.pose_landmarker.detect_async(mp_image, timestamp_ms)
-            if "hand" in self.active_tasks and self.hand_landmarker:
-                self.hand_landmarker.detect_async(mp_image, timestamp_ms)
-            if "object" in self.active_tasks and self.object_detector:
-                self.object_detector.detect_async(mp_image, timestamp_ms)
+            # 모든 활성 Task에서 비동기 처리
+            processing_tasks = []
+            
+            for task_type, task in self.active_tasks.items():
+                if self.task_health.get(task_type, False):
+                    try:
+                        if hasattr(task, 'detect_async'):
+                            task.detect_async(mp_image, timestamp_ms)
+                        elif hasattr(task, 'recognize_async'):
+                            task.recognize_async(mp_image, timestamp_ms)
+                    except Exception as e:
+                        logger.warning(f"{task_type.value} 처리 오류: {e}")
+                        self.task_health[task_type] = False
+            
+            # 처리 시간 기록
+            processing_time = time.time() - start_time
+            self.frame_processing_times.append(processing_time)
+            
+            # deque가 자동으로 maxlen 관리하므로 수동 제거 불필요
+            
         except Exception as e:
-            logger.error(f"MediaPipe 태스크 실행 중 오류: {e}")
+            logger.error(f"프레임 처리 오류: {e}")
+        
+        return self.latest_results.copy()
 
-    def get_latest_results(self):
+    def _calculate_fps(self):
+        """FPS 계산"""
+        self.fps_counter += 1
+        if self.fps_counter % 30 == 0:  # 30프레임마다 계산
+            current_time = time.time()
+            self.current_fps = 30 / (current_time - self.fps_start_time)
+            self.fps_start_time = current_time
+
+    def get_performance_stats(self) -> Dict[str, Any]:
+        """성능 통계 반환"""
+        avg_processing_time = np.mean(self.frame_processing_times) if self.frame_processing_times else 0
+        
         return {
-            "face": self.last_face_result,
-            "pose": self.last_pose_result,
-            "hand": self.last_hand_result,
-            "object": self.last_object_result,
+            "fps": self.current_fps,
+            "avg_processing_time_ms": avg_processing_time * 1000,
+            "active_tasks": len(self.active_tasks),
+            "healthy_tasks": sum(self.task_health.values()),
+            "task_health": self.task_health.copy(),
+            "queue_size": self.result_queue.qsize()
         }
 
-    def close(self):
-        logger.info("MediaPipe Task 리소스 정리 시작...")
+    def get_latest_results(self) -> Dict[str, Any]:
+        """최신 결과 반환 (호환성을 위해 문자열 키 사용)"""
+        return {
+            "face": self.latest_results.get(TaskType.FACE_LANDMARKER),
+            "pose": self.latest_results.get(TaskType.POSE_LANDMARKER),
+            "hand": self.latest_results.get(TaskType.HAND_LANDMARKER),
+            "gesture": self.latest_results.get(TaskType.GESTURE_RECOGNIZER),
+            "object": self.latest_results.get(TaskType.OBJECT_DETECTOR),
+            "holistic": self.latest_results.get(TaskType.HOLISTIC_LANDMARKER)
+        }
+
+    async def close(self):
+        """리소스 정리"""
+        logger.info("MediaPipe Manager 정리 시작...")
         
-        # Set shutdown flag first
-        self._shutdown_requested = True
+        self.running = False
         
-        # Send shutdown signal to queue
-        try:
-            self.result_queue.put(('shutdown', None, None))
-        except Exception as e:
-            logger.warning(f"Shutdown signal 전송 실패: {e}")
-        
-        # Wait for callback thread to finish with timeout
+        # 콜백 스레드 종료
         if self.callback_thread and self.callback_thread.is_alive():
-            self.callback_thread.join(timeout=2.0)
-            if self.callback_thread.is_alive():
-                logger.warning("Callback thread가 정상적으로 종료되지 않았습니다.")
+            self.result_queue.put(("shutdown", None, None))
+            self.callback_thread.join(timeout=5.0)
         
-        # Close MediaPipe tasks
-        for task in [self.face_landmarker, self.pose_landmarker, self.hand_landmarker, self.object_detector]:
-            if task:
-                try:
+        # 모든 Task 정리
+        for task_type, task in self.active_tasks.items():
+            try:
+                if hasattr(task, 'close'):
                     task.close()
-                except Exception as e:
-                    logger.warning(f"MediaPipe task 종료 중 오류: {e}")
+                logger.info(f"✅ {task_type.value} 정리 완료")
+            except Exception as e:
+                logger.warning(f"❌ {task_type.value} 정리 오류: {e}")
         
-        logger.info("MediaPipe Task 리소스 정리 완료.")
+        self.active_tasks.clear()
+        self.task_health.clear()
+        self.latest_results.clear()
+        
+        logger.info("MediaPipe Manager 정리 완료")
+
+    # 레거시 호환성 메소드들
+    def run_tasks(self, frame):
+        """레거시 호환성을 위한 동기 인터페이스"""
+        return asyncio.run(self.process_frame(frame))
+
+    @property 
+    def current_fps(self):
+        return getattr(self, '_current_fps', 0.0)
+    
+    @current_fps.setter
+    def current_fps(self, value):
+        self._current_fps = value
